Exit on error

+++ run_tests_in_containers.sh
@@ -56,7 +56,12 @@ sleep 5
 # Upload der to Solr
 echo "Upload public der key to Solr"
 docker exec -d $SOLR_CONT_NAME \
-    ls -l /opt/tests/$SIGNING_PUBLIC_KEY_DER && bin/solr package add-key /opt/tests/$SIGNING_PUBLIC_KEY_DER
+    "ls -l /opt/tests/$SIGNING_PUBLIC_KEY_DER && bin/solr package add-key /opt/tests/$SIGNING_PUBLIC_KEY_DER"
+
+if [ "$?" -gt "0" ]; then
+  echo "Could not add public DER key to solr cloud"
+  exit 1
+fi
 
 # For atlas-web-bulk-cli application context
 docker run --rm --net $DOCKER_NET --name $POSTGRES_HOST \
